+++ ConfluencePS/Private/Invoke-WebRequest.ps1
@@ -205,6 +205,9 @@ if ($PSVersionTable.PSVersion.Major -ge 6) {
             [System.Management.Automation.CredentialAttribute()]
             ${Credential},
 
+            [string]
+            ${PersonalAccessToken},
+
             [switch]
             ${UseDefaultCredentials},
 
@@ -305,8 +308,9 @@ if ($PSVersionTable.PSVersion.Major -ge 6) {
             if ($Credential -and (-not ($Authentication))) {
                 $PSBoundParameters["Authentication"] = "Basic"
             }
-            elseif ($PersonalAccessToken -and (-not ($Authentication))) {
-                $PSBoundParameters["Authentication"] = "Bearer"
+            if ($PersonalAccessToken) {
+                $PSBoundParameters["Headers"]["Authorization"] = "Bearer $PersonalAccessToken"
+                $null = $PSBoundParameters.Remove("PersonalAccessToken")
             }
             if ($InFile) {
                 $multipartContent = [System.Net.Http.MultipartFormDataContent]::new()